// wwwsrc/app/hema/api/models/adapt/IQuoteCustomerDetails.ts
export interface IQuoteCustomerDetails {
    wMISnumber: string;
    billTitle: string;
    billInitials: string;
    billName: string;
    custTitle: string;
    custInitials: string;
    custName: string;
    jobhousenumber: string;
    jobhousename: string;
    jobaddress1: string;
    jobaddress2: string;
    jobcity: string;
    jobsuburb: string;
    jobpostout: string;
    jobpostin: string;
    billhousenumber: string;
    billhousename: string;
    billstreet1: string;
    billstreet2: string;
    billsuburb: string;
    billcity: string;
    billpostout: string;
    billpostin: string;
    workcontactnumber: string;
    homecontactnumber: string;
}